+++ game/components/c_fire.h
@@ -10,7 +10,9 @@ namespace Verse::Component
 {
     struct Fire {
         Rect2 transform;
-        Vec2 offset = Vec2(0, 0);
+        Vec2 offset;
+        Vec2 initial_offset;
+        std::vector<int> vertical_offsets = {};
         
         ui32 flame_tex, p_tex;
         ui8* p_data;